refactor: seperate reconciliation logic
